[nvim] Enable CopilotChat plugin
This commit enables the CopilotChat plugin in the Neovim configuration.
The plugin was previously commented out in the configuration files. The
changes include uncommenting the relevant lines in the `flake.nix` and
`default.nix` files, and adding new key mappings for the plugin in the
`init.lua` file.

The new key mappings include a quick chat command mapped to `<leader>ccq`
and a command to open the Copilot Chat mapped to `<leader>cp`.